requirements.txt
astroid==2.3.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.6
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
pylint==2.4.2
pytz==2019.3
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wincertstore==0.2
wrapt==1.11.2
